requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic[email]>=2.5.0
websockets>=12.0
cryptography>=41.0.7
bcrypt>=4.1.1
email-validator>=2.1.0
